+++ src/core/lib/decorators/index.ts
@@ -1,0 +1,5 @@
+export * from './class';
+export * from './errorMiddleware';
+export * from './method';
+export * from './middleware';
+export * from './wrapper';
